+++ example/packages/manifest/src/index.ts
@@ -1,9 +1,10 @@
+import type { OpenAISchema } from '@tailor-cms/cek-common';
+
 import type {
   DataInitializer,
   ElementData,
   ElementManifest,
 } from './interfaces';
-import ai from './ai';
 
 // Element unique id within the target system (e.g. Tailor)
 export const type = 'ACME_TCE_COUNTER';
@@ -37,7 +38,29 @@ export const mocks = {
   ],
 };
 
-export { ai };
+export const ai = {
+  Schema: {
+    type: 'json_schema',
+    name: 'ce_counter',
+    schema: {
+      type: 'object',
+      properties: {
+        count: { type: 'number' },
+        description: { type: 'string' },
+      },
+      required: ['count', 'description'],
+      additionalProperties: false,
+    },
+  } as OpenAISchema,
+  getPrompt: () => `
+    Generate counter content element as an object with the following
+    properties: { "description": "", "count": 0 }.
+    Where:
+    - 'description' is the counter description
+    - 'count' is the number representing counter value
+  `,
+  processResponse: (val: any) => val,
+};
 
 const manifest: ElementManifest = {
   type,
@@ -45,8 +68,8 @@ const manifest: ElementManifest = {
   name,
   initState,
   ui,
-  ai,
   mocks,
+  ai,
 };
 
 export default manifest;