+++ packages/runtime/tce-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tailor-cms/tce-server-runtime
 
+## 1.2.5
+
+### Patch Changes
+
+- Update content generation handling.
+
+## 1.2.4
+
+### Patch Changes
+
+- Remove request timeout, fix focus on content generate.
+
 ## 1.2.3
 
 ### Patch Changes